+++ src/EMRG/Web/Areas/DepartmentAdmin/Pages/Courses/Edit.cshtml.cs
@@ -53,7 +53,7 @@ namespace Web.Areas.DepartmentAdmin.Pages.Courses
         public async Task<IActionResult> OnPostAsync()
         {
             var user = await userManager.GetUserAsync(User);
-            Course.DepartmentId = user.DepartmentId;
+            Course.DepartmentId = (int)user.DepartmentId;
             if (!ModelState.IsValid)
             {
                 return Page();
